+++ tests/Rule/NoReuseTest.php
@@ -53,10 +53,10 @@ final class NoReuseTest extends TestCase
      */
     public function testCanConstructRuleWithCountConstraint(): void
     {
-        $rule = new NoReuse($this->hashFunction, 5, 1);
+        $rule = new NoReuse($this->hashFunction, 5, 0);
 
         // Force generation of code coverage
-        $ruleConstruct = new NoReuse($this->hashFunction, 5, 1);
+        $ruleConstruct = new NoReuse($this->hashFunction, 5, 0);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -91,7 +91,7 @@ final class NoReuseTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new NoReuse($this->hashFunction, 0, 1);
+        $rule = new NoReuse($this->hashFunction, 0, 0);
     }
 
     /**
@@ -101,17 +101,19 @@ final class NoReuseTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new NoReuse($this->hashFunction, -10, 1);
+        $rule = new NoReuse($this->hashFunction, -10, 0);
     }
 
     /**
      * @covers ::__construct
      */
-    public function testCannotConstructRuleWithFirstConstraintEqualToZero(): void
+    public function testCanConstructRuleWithFirstConstraintEqualToZero(): void
     {
-        $this->expectException(InvalidArgumentException::class);
-
         $rule = new NoReuse($this->hashFunction, null, 0);
+
+        // Force generation of code coverage
+        $ruleConstruct = new NoReuse($this->hashFunction, null, 0);
+        self::assertEquals($rule, $ruleConstruct);
     }
 
     /**
@@ -125,26 +127,6 @@ final class NoReuseTest extends TestCase
     }
 
     /**
-     * @covers ::getCount
-     */
-    public function testCanGetCountConstraint(): void
-    {
-        $rule = new NoReuse($this->hashFunction, 5, 10);
-
-        self::assertSame(5, $rule->getCount());
-    }
-
-    /**
-     * @covers ::getFirst
-     */
-    public function testCanGetFirstConstraint(): void
-    {
-        $rule = new NoReuse($this->hashFunction, 5, 10);
-
-        self::assertSame(10, $rule->getFirst());
-    }
-
-    /**
      * @covers ::getHashFunction
      */
     public function testCanGetHashFunction(): void
@@ -155,11 +137,39 @@ final class NoReuseTest extends TestCase
     }
 
     /**
+     * @covers ::addConstraint
+     */
+    public function testCanAddConstraint(): void
+    {
+        $rule = new NoReuse($this->hashFunction, 5, 5, 1);
+        $rule->addConstraint(10, 10, 1);
+
+        // Force generation of code coverage
+        $ruleConstruct = new NoReuse($this->hashFunction, 5, 5, 1);
+        $ruleConstruct->addConstraint(10, 10, 1);
+        self::assertEquals($rule, $ruleConstruct);
+    }
+
+    /**
+     * @covers ::addConstraint
+     */
+    public function testConstraintsAreOrdered(): void
+    {
+        $rule = new NoReuse($this->hashFunction, 5, 5, 1);
+        $rule->addConstraint(10, 10, 2);
+
+        // Force generation of code coverage
+        $ruleConstruct = new NoReuse($this->hashFunction, 10, 10, 2);
+        $ruleConstruct->addConstraint(5, 5, 1);
+        self::assertEquals($rule, $ruleConstruct);
+    }
+
+    /**
      * @covers ::test
      */
     public function testRuleIsSatisfiedWhenPasswordIsString(): void
     {
-        $rule = new NoReuse($this->hashFunction, null, 1);
+        $rule = new NoReuse($this->hashFunction, null, 0);
 
         self::assertTrue($rule->test('foobar'));
     }
@@ -169,7 +179,7 @@ final class NoReuseTest extends TestCase
      */
     public function testCountConstraintCanBeSatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, 3, 1);
+        $rule = new NoReuse($this->hashFunction, 3, 0);
 
         self::assertTrue($rule->test($this->password));
     }
@@ -179,7 +189,7 @@ final class NoReuseTest extends TestCase
      */
     public function testCountConstraintCanBeUnsatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, 4, 1);
+        $rule = new NoReuse($this->hashFunction, 4, 0);
 
         self::assertFalse($rule->test($this->password));
     }
@@ -189,7 +199,7 @@ final class NoReuseTest extends TestCase
      */
     public function testFirstConstraintCanBeSatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, null, 5);
+        $rule = new NoReuse($this->hashFunction, null, 4);
 
         self::assertTrue($rule->test($this->password));
     }
@@ -199,7 +209,7 @@ final class NoReuseTest extends TestCase
      */
     public function testFirstConstraintCanBeUnsatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, null, 4);
+        $rule = new NoReuse($this->hashFunction, null, 3);
 
         self::assertFalse($rule->test($this->password));
     }
@@ -209,12 +219,12 @@ final class NoReuseTest extends TestCase
      */
     public function testEnforceDoesNotThrowExceptionWhenRuleIsSatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, 1, 1);
+        $rule = new NoReuse($this->hashFunction, 1, 0);
 
         $rule->enforce($this->password);
 
         // Force generation of code coverage
-        $ruleConstruct = new NoReuse($this->hashFunction, 1, 1);
+        $ruleConstruct = new NoReuse($this->hashFunction, 1, 0);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -223,7 +233,7 @@ final class NoReuseTest extends TestCase
      */
     public function testEnforceThrowsExceptionWhenRuleIsNotSatisfied(): void
     {
-        $rule = new NoReuse($this->hashFunction, null, 1);
+        $rule = new NoReuse($this->hashFunction, null, 0);
 
         $this->expectException(RuleException::class);